feat(settings): Updated Settings Page (#32)
* Added avatars to settings page

* Updated settings page
